feat(io): implement HDF5/XDMF particle exporter and fix FMM tree bugs
- Implement HDF5Exporter class to write per-frame particle trajectories into binary HDF5 files and master XDMF index for ParaView
- Add Apptainer container definition (Apptainer.def) with C++20, OpenMPI, and parallel HDF5 support
- Update main driver (src/main2.cpp) to export particle states during FMM simulation steps
- Fix mass center accumulation in Node::calculateMC by replacing assignment with addition (+=)
- Fix 3D spherical harmonics moment accumulation in MultipoleSet<3>::operator+= and fix loop increment over m
- Fix out-of-bounds evaluation in NodeI::collect_multipoles_to_locals interaction list loop
- Update Morton Z-order encoding/decoding functions to use size_t for extended bit precision

+++ src/system_updaters/tree_components/morton_code.cpp
@@ -6,7 +6,7 @@
 #include "morton_code.hpp"
 #include <cstring>
 
-unsigned int Morton::encode(const unsigned int *cartesian_coordinates, unsigned int dim){
+size_t Morton::encode(const unsigned int *cartesian_coordinates, unsigned int dim){
   unsigned int code = 0;
   unsigned int temp_cartesian_coordinates[dim];
   std::memcpy(temp_cartesian_coordinates, cartesian_coordinates, sizeof(unsigned int) * dim);
@@ -21,7 +21,7 @@ unsigned int Morton::encode(const unsigned int *cartesian_coordinates, unsigned 
   return code;
 }
 
-void Morton::decode(unsigned int *cartesian_coordinates, unsigned int dim, unsigned code){
+void Morton::decode(unsigned int *cartesian_coordinates, unsigned int dim, size_t code){
   for (unsigned int i = 0; i < dim; ++i)
     cartesian_coordinates[i] = 0;
