+++ src/system_updaters/tree_components/morton_code.hpp
@@ -1,4 +1,5 @@
 #pragma once
+#include <cstddef>
 
 /**
  * @file morton_code.hpp
@@ -17,7 +18,7 @@ class Morton {
      * @param dim Spatial dimension \f$ d \f$.
      * @return Encoded Morton Z-order integer code.
      */
-    static unsigned int encode(const unsigned int* cartesian_coordinates, unsigned int dim);
+    static size_t encode(const unsigned int* cartesian_coordinates, unsigned int dim);
 
     /**
      * @brief Decodes a 1D Morton code back into multi-dimensional integer Cartesian coordinates.
@@ -25,5 +26,5 @@ class Morton {
      * @param dim Spatial dimension \f$ d \f$.
      * @param code Morton Z-order integer code to decode.
      */
-    static void decode(unsigned int* cartesian_coordinates, unsigned int dim, unsigned code);
+    static void decode(unsigned int* cartesian_coordinates, unsigned int dim, size_t code);
 };
